# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/assessment-test": {
      "post": {
        "operationId": "AssessmentTestController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssessmentTestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentTestDto"
                }
              }
            }
          }
        },
        "tags": [
          "Assessment Tests"
        ]
      },
      "get": {
        "operationId": "AssessmentTestController_fetch",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssessmentTestDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Assessment Tests"
        ]
      },
      "patch": {
        "operationId": "AssessmentTestController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssessmentTestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentTestDto"
                }
              }
            }
          }
        },
        "tags": [
          "Assessment Tests"
        ]
      },
      "delete": {
        "operationId": "AssessmentTestController_delete",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Assessment Tests"
        ]
      }
    },
    "/assessment-test/user-asssessments": {
      "get": {
        "operationId": "AssessmentTestController_fetchUsersAssessments",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserAssessmentTestDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Assessment Tests"
        ]
      }
    },
    "/assessment-test/admin-user-asssessments/{id}": {
      "get": {
        "operationId": "AssessmentTestController_fetchAdminUserAssessments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserAssessmentTestDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Assessment Tests"
        ]
      }
    },
    "/assessment-test/user-subjects-eligibility": {
      "get": {
        "operationId": "AssessmentTestController_fetchUserSubjectsEligibility",
        "parameters": [
          {
            "name": "subjects",
            "required": true,
            "in": "query",
            "description": "CSV list of subjects, e.g. ANGULAR,NESTJS,RXJS",
            "schema": {
              "example": "ANGULAR,NESTJS",
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "ANGULAR",
                  "NESTJS",
                  "RXJS"
                ]
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSubjectEligibilityDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Assessment Tests"
        ]
      }
    },
    "/assessment-test/admin-user-subjects-eligibility/{id}": {
      "get": {
        "operationId": "AssessmentTestController_fetchAdminUserSubjectsEligibility",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subjects",
            "required": true,
            "in": "query",
            "description": "CSV list of subjects, e.g. ANGULAR,NESTJS,RXJS",
            "schema": {
              "example": "ANGULAR,NESTJS",
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "ANGULAR",
                  "NESTJS",
                  "RXJS"
                ]
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSubjectEligibilityDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Assessment Tests"
        ]
      }
    },
    "/assessment-test/start-test": {
      "post": {
        "operationId": "AssessmentTestController_startTest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartTestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAssessmentTestDto"
                }
              }
            }
          }
        },
        "tags": [
          "Assessment Tests"
        ]
      }
    },
    "/assessment-test/submit-test": {
      "post": {
        "operationId": "AssessmentTestController_submitTest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitTestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAssessmentTestDto"
                }
              }
            }
          }
        },
        "tags": [
          "Assessment Tests"
        ]
      }
    },
    "/assessment-test/attempts/{id}/questions": {
      "get": {
        "operationId": "AssessmentTestController_questions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptQuestionsDto"
                }
              }
            }
          }
        },
        "tags": [
          "Assessment Tests"
        ]
      }
    },
    "/assessment-test/{id}": {
      "delete": {
        "operationId": "AssessmentTestController_deleteById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Assessment Tests"
        ]
      },
      "get": {
        "operationId": "AssessmentTestController_fetchAssessmentTest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentTestDto"
                }
              }
            }
          }
        },
        "tags": [
          "Assessment Tests"
        ]
      }
    }
  },
  "info": {
    "title": "NGX service-nestjs-assessment-test API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "ChoiceInputDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          }
        },
        "required": [
          "value"
        ]
      },
      "QuestionInputDto": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string"
          },
          "choices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChoiceInputDto"
            }
          },
          "answer": {
            "type": "string"
          },
          "correctResponse": {
            "type": "string"
          },
          "incorrectResponse": {
            "type": "string"
          }
        },
        "required": [
          "question",
          "choices",
          "answer",
          "correctResponse",
          "incorrectResponse"
        ]
      },
      "CreateAssessmentTestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "subject": {
            "type": "string",
            "enum": [
              "ANGULAR",
              "NESTJS",
              "RXJS"
            ]
          },
          "level": {
            "type": "integer",
            "minimum": 1
          },
          "testQuestions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionInputDto"
            }
          }
        },
        "required": [
          "name",
          "subject",
          "level",
          "testQuestions"
        ]
      },
      "TestChoiceDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          }
        },
        "required": [
          "value"
        ]
      },
      "TestQuestionDto": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string"
          },
          "choices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TestChoiceDto"
            }
          },
          "answer": {
            "type": "string"
          },
          "correctResponse": {
            "type": "string"
          },
          "incorrectResponse": {
            "type": "string"
          }
        },
        "required": [
          "question",
          "choices",
          "answer",
          "correctResponse",
          "incorrectResponse"
        ]
      },
      "AssessmentTestDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "subject": {
            "type": "string",
            "enum": [
              "ANGULAR",
              "NESTJS",
              "RXJS"
            ]
          },
          "level": {
            "type": "number"
          },
          "lastUpdated": {
            "type": "string",
            "format": "date-time"
          },
          "testQuestions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TestQuestionDto"
            }
          },
          "__v": {
            "type": "number"
          }
        },
        "required": [
          "_id",
          "name",
          "subject",
          "level",
          "lastUpdated",
          "testQuestions",
          "__v"
        ]
      },
      "UpdateAssessmentTestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "subject": {
            "type": "string",
            "enum": [
              "ANGULAR",
              "NESTJS",
              "RXJS"
            ]
          },
          "level": {
            "type": "integer",
            "minimum": 1
          },
          "testQuestions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionInputDto"
            }
          },
          "_id": {
            "type": "string"
          },
          "__v": {
            "type": "number",
            "description": "Version returned by the definition read; prevents lost updates."
          }
        },
        "required": [
          "name",
          "subject",
          "level",
          "testQuestions",
          "_id",
          "__v"
        ]
      },
      "UserAssessmentTestDto": {
        "type": "object",
        "properties": {
          "assessmentTestId": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "description": "Canonical owner; legacy records may only have uuid"
          },
          "_id": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          },
          "testName": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "subject": {
            "type": "string",
            "enum": [
              "ANGULAR",
              "NESTJS",
              "RXJS"
            ]
          },
          "userAnswers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "passed": {
            "type": "boolean"
          },
          "completed": {
            "type": "boolean"
          },
          "lastUpdated": {
            "type": "string",
            "format": "date-time"
          },
          "__v": {
            "type": "number"
          }
        },
        "required": [
          "assessmentTestId",
          "_id",
          "uuid",
          "testName",
          "score",
          "subject",
          "userAnswers",
          "passed",
          "completed",
          "lastUpdated",
          "__v"
        ]
      },
      "UserSubjectEligibilityDto": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "enum": [
              "ANGULAR",
              "NESTJS",
              "RXJS"
            ]
          },
          "levelCount": {
            "type": "number"
          },
          "totalCount": {
            "type": "number"
          },
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "subject",
          "levelCount",
          "totalCount",
          "enabled"
        ]
      },
      "StartTestDto": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "enum": [
              "ANGULAR",
              "NESTJS",
              "RXJS"
            ]
          }
        },
        "required": [
          "subject"
        ]
      },
      "SubmitTestDto": {
        "type": "object",
        "properties": {
          "testId": {
            "type": "string"
          },
          "answers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "testId",
          "answers"
        ]
      },
      "LearnerQuestionDto": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string"
          },
          "choices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChoiceInputDto"
            }
          }
        },
        "required": [
          "question",
          "choices"
        ]
      },
      "AttemptQuestionsDto": {
        "type": "object",
        "properties": {
          "testId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "testQuestions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LearnerQuestionDto"
            }
          }
        },
        "required": [
          "testId",
          "name",
          "testQuestions"
        ]
      }
    }
  }
}